+++ metric/requests_inflight.go
@@ -13,7 +13,9 @@ const (
 	metricDescRequestInFlight = "Measures the number of requests currently being processed by the server."
 )
 
-// [RequestInFlight] is a metrics recorder for recording the number of requests in flight.
+// NewRequestInFlight records the number of requests currently being processed by the server.
+//
+// Deprecated: use NewServerActiveRequests instead.
 func NewRequestInFlight(cfg BaseConfig) func(next http.Handler) http.Handler {
 	// init metric, here we are using counter for capturing request in flight
 	counter, err := cfg.Meter.Int64UpDownCounter(
@@ -27,6 +29,11 @@ func NewRequestInFlight(cfg BaseConfig) func(next http.Handler) http.Handler {
 
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			if !cfg.ShouldRecord(r) {
+				next.ServeHTTP(w, r)
+				return
+			}
+
 			// define metric attributes
 			attrs := otelmetric.WithAttributes(cfg.AttributesFunc(r)...)